# E2E Performance - printing latency numbers for volume lifecycle operations