--- a/test/e2e/storage/vsphere_volume_perf.go
+++ b/test/e2e/storage/vsphere_volume_perf.go
@@ -26,22 +26,45 @@ import (
 	. "github.com/onsi/gomega"
 	"k8s.io/api/core/v1"
 	storageV1 "k8s.io/api/storage/v1"
-	k8stype "k8s.io/apimachinery/pkg/types"
+	"k8s.io/apimachinery/pkg/types"
 	clientset "k8s.io/client-go/kubernetes"
 	vsphere "k8s.io/kubernetes/pkg/cloudprovider/providers/vsphere"
 	"k8s.io/kubernetes/test/e2e/framework"
 )
 
+/* This test calculates latency numbers for volume lifecycle operations
+1. Create 4 type of storage classes
+2. Read the total number of volumes to be created and volumes per pod
+3. Create total PVCs (number of volumes)
+4. Create Pods with attached volumes per pod
+5. Verify access to the volumes
+6. Delete pods and wait for volumes to detach
+7. Delete the PVCs
+*/
+const (
+	NodeLabelKey              = "vsphere_e2e_label"
+	SCSIUnitsAvailablePerNode = 55
+	NumOps                    = 4
+)
+
+// NodeSelector holds
+type NodeSelector struct {
+	labelKey   string
+	labelValue string
+}
+
 var _ = SIGDescribe("vcp-performance", func() {
 	f := framework.NewDefaultFramework("vcp-performance")
 
 	var (
-		client    clientset.Interface
-		namespace string
-		//iterations int
+		client           clientset.Interface
+		namespace        string
+		nodeSelectorList []*NodeSelector
+		volumeCount      int
 	)
 
 	BeforeEach(func() {
+		var err error
 		framework.SkipUnlessProviderIs("vsphere")
 		client = f.ClientSet
 		namespace = f.Namespace.Name
@@ -50,11 +73,21 @@ var _ = SIGDescribe("vcp-performance", func() {
 		Expect(os.Getenv("VSPHERE_DATASTORE")).NotTo(BeEmpty(), "env VSPHERE_DATASTORE is not set")
 		Expect(os.Getenv("VCP_PERF_VOLUME_PER_POD")).NotTo(BeEmpty(), "env VCP_PERF_VOLUME_PER_POD is not set")
 		Expect(os.Getenv("VCP_PERF_ITERATIONS")).NotTo(BeEmpty(), "env VCP_PERF_ITERATIONS is not set")
+		Expect(os.Getenv("VCP_PERF_VOLUME_COUNT")).NotTo(BeEmpty(), "env VCP_PERF_VOLUME_COUNT is not set")
+
+		// Verify volume count specified by the user can be satisfied
+		volumeCount, err = strconv.Atoi(os.Getenv("VCP_PERF_VOLUME_COUNT"))
+		Expect(err).NotTo(HaveOccurred(), "Error Parsing VCP_PERF_VOLUME_COUNT")
 
 		nodes := framework.GetReadySchedulableNodesOrDie(client)
 		if len(nodes.Items) < 2 {
 			framework.Skipf("Requires at least %d nodes (not %d)", 2, len(nodes.Items))
 		}
+		if volumeCount > SCSIUnitsAvailablePerNode*len(nodes.Items) {
+			framework.Skipf("Cannot attach %d volumes to %d nodes. Maximum volumes that can be attached on %d nodes is %d", volumeCount, len(nodes.Items), len(nodes.Items), SCSIUnitsAvailablePerNode*len(nodes.Items))
+		}
+		nodeSelectorList = createNodeLabels(client, namespace, nodes)
+
 	})
 
 	It("vcp performance tests", func() {
@@ -65,14 +98,20 @@ var _ = SIGDescribe("vcp-performance", func() {
 		iterations, err := strconv.Atoi(os.Getenv("VCP_PERF_ITERATIONS"))
 		Expect(err).NotTo(HaveOccurred(), "Error Parsing VCP_PERF_ITERATIONS")
 
+		var sumLatency [NumOps]int64
 		for i := 0; i < iterations; i++ {
-			latency := invokeVolumeLifeCyclePerformance(f, client, namespace, scList, volumesPerPod)
-			framework.Logf("Performance numbers for iteration %d", i)
-			framework.Logf("Creating PVCs and waiting for bound phase: %v microseconds", latency[0])
-			framework.Logf("Creating Pod and verifying attached status: %v microseconds", latency[1])
-			framework.Logf("Deleting Pod and waiting for disk to be detached: %v microseconds", latency[2])
-			framework.Logf("Deleting the PVCs: %v microseconds", latency[3])
+			latency := invokeVolumeLifeCyclePerformance(f, client, namespace, scList, volumesPerPod, volumeCount, nodeSelectorList)
+			for i, val := range latency {
+				sumLatency[i] += val
+			}
 		}
+
+		framework.Logf("Average latency for below operations")
+		framework.Logf("Creating PVCs and waiting for bound phase: %v microseconds", sumLatency[0]/NumOps)
+		framework.Logf("Creating Pod: %v microseconds", sumLatency[1]/NumOps)
+		framework.Logf("Deleting Pod and waiting for disk to be detached: %v microseconds", sumLatency[2]/NumOps)
+		framework.Logf("Deleting the PVCs: %v microseconds", sumLatency[3]/NumOps)
+
 		for _, sc := range scList {
 			client.StorageV1().StorageClasses().Delete(sc.Name, nil)
 		}
@@ -124,56 +163,100 @@ func getTestStorageClasses(client clientset.Interface) []*storageV1.StorageClass
 }
 
 // invokeVolumeLifeCyclePerformance peforms full volume life cycle management and records latency for each operation
-func invokeVolumeLifeCyclePerformance(f *framework.Framework, client clientset.Interface, namespace string, sc []*storageV1.StorageClass, volumesPerPod int) []int64 {
+func invokeVolumeLifeCyclePerformance(f *framework.Framework, client clientset.Interface, namespace string, sc []*storageV1.StorageClass, volumesPerPod int, volumeCount int, nodeSelectorList []*NodeSelector) []int64 {
 	const timeFraction = 1000 // Show results in microseconds
-	var latency []int64
+	var (
+		latency       []int64
+		totalpvclaims [][]*v1.PersistentVolumeClaim
+		totalpvs      [][]*v1.PersistentVolume
+		totalpods     []*v1.Pod
+	)
+	nodeVolumeMap := make(map[types.NodeName][]string)
 
-	By(fmt.Sprintf("Creating %v PVCs per pod ", volumesPerPod))
-	var pvclaims []*v1.PersistentVolumeClaim
+	numPods := volumeCount / volumesPerPod
+
+	By(fmt.Sprintf("Creating %v PVCs", volumeCount))
 	start := time.Now()
-	for i := 0; i < volumesPerPod; i++ {
-		pvclaim, err := framework.CreatePVC(client, namespace, getVSphereClaimSpecWithStorageClassAnnotation(namespace, sc[i%len(sc)]))
-		Expect(err).NotTo(HaveOccurred())
-		pvclaims = append(pvclaims, pvclaim)
+	for i := 0; i < numPods; i++ {
+		var pvclaims []*v1.PersistentVolumeClaim
+		for j := 0; j < volumesPerPod; j++ {
+			currsc := sc[((i*numPods)+j)%len(sc)]
+			pvclaim, err := framework.CreatePVC(client, namespace, getVSphereClaimSpecWithStorageClassAnnotation(namespace, currsc))
+			Expect(err).NotTo(HaveOccurred())
+			pvclaims = append(pvclaims, pvclaim)
+		}
+		totalpvclaims = append(totalpvclaims, pvclaims)
 	}
-
-	persistentvolumes, err := framework.WaitForPVClaimBoundPhase(client, pvclaims, framework.ClaimProvisionTimeout)
-	Expect(err).NotTo(HaveOccurred())
+	for _, pvclaims := range totalpvclaims {
+		persistentvolumes, err := framework.WaitForPVClaimBoundPhase(client, pvclaims, framework.ClaimProvisionTimeout)
+		Expect(err).NotTo(HaveOccurred())
+		totalpvs = append(totalpvs, persistentvolumes)
+	}
 	elapsed := time.Since(start)
 	latency = append(latency, elapsed.Nanoseconds()/timeFraction)
 
-	By("Creating pod to attach PVs to the node and verifying access")
+	By("Creating pod to attach PVs to the node")
 	start = time.Now()
-	pod, err := framework.CreatePod(client, namespace, pvclaims, false, "")
-	Expect(err).NotTo(HaveOccurred())
-
-	vsp, err := vsphere.GetVSphere()
-	Expect(err).NotTo(HaveOccurred())
-
-	verifyVSphereVolumesAccessible(pod, persistentvolumes, vsp)
+	for i, pvclaims := range totalpvclaims {
+		nodeSelector := nodeSelectorList[i%len(nodeSelectorList)]
+		pod, err := framework.CreatePod(client, namespace, map[string]string{nodeSelector.labelKey: nodeSelector.labelValue}, pvclaims, false, "")
+		Expect(err).NotTo(HaveOccurred())
+		totalpods = append(totalpods, pod)
+	}
 	elapsed = time.Since(start)
 	latency = append(latency, elapsed.Nanoseconds()/timeFraction)
 
-	By("Deleting pod and waiting for volumes to be detached from the node")
-	start = time.Now()
-	err = framework.DeletePodWithWait(f, client, pod)
+	// Verify access to the volumes
+	vsp, err := vsphere.GetVSphere()
 	Expect(err).NotTo(HaveOccurred())
 
-	for _, pv := range persistentvolumes {
-		err = waitForVSphereDiskToDetach(vsp, pv.Spec.VsphereVolume.VolumePath, k8stype.NodeName(pod.Spec.NodeName))
+	for i, pod := range totalpods {
+		verifyVSphereVolumesAccessible(pod, totalpvs[i], vsp)
+	}
+
+	By("Deleting pods")
+	start = time.Now()
+	for _, pod := range totalpods {
+		err = framework.DeletePodWithWait(f, client, pod)
 		Expect(err).NotTo(HaveOccurred())
 	}
 	elapsed = time.Since(start)
 	latency = append(latency, elapsed.Nanoseconds()/timeFraction)
 
+	for i, pod := range totalpods {
+		for _, pv := range totalpvs[i] {
+			nodeName := types.NodeName(pod.Spec.NodeName)
+			nodeVolumeMap[nodeName] = append(nodeVolumeMap[nodeName], pv.Spec.VsphereVolume.VolumePath)
+		}
+	}
+
+	err = waitForVSphereDisksToDetach(vsp, nodeVolumeMap)
+	Expect(err).NotTo(HaveOccurred())
+
 	By("Deleting the PVCs")
 	start = time.Now()
-	for _, pvc := range pvclaims {
-		err = framework.DeletePersistentVolumeClaim(client, pvc.Name, namespace)
-		Expect(err).NotTo(HaveOccurred())
+	for _, pvclaims := range totalpvclaims {
+		for _, pvc := range pvclaims {
+			err = framework.DeletePersistentVolumeClaim(client, pvc.Name, namespace)
+			Expect(err).NotTo(HaveOccurred())
+		}
 	}
 	elapsed = time.Since(start)
 	latency = append(latency, elapsed.Nanoseconds()/timeFraction)
 
 	return latency
+}
+
+func createNodeLabels(client clientset.Interface, namespace string, nodes *v1.NodeList) []*NodeSelector {
+	var nodeSelectorList []*NodeSelector
+	for i, node := range nodes.Items {
+		labelVal := "vsphere_e2e_" + strconv.Itoa(i)
+		nodeSelector := &NodeSelector{
+			labelKey:   NodeLabelKey,
+			labelValue: labelVal,
+		}
+		nodeSelectorList = append(nodeSelectorList, nodeSelector)
+		framework.AddOrUpdateLabelOnNode(client, node.Name, NodeLabelKey, labelVal)
+	}
+	return nodeSelectorList
 }